requirements.txt
asgiref==3.8.1
Django==5.1
django-phonenumber-field==8.0.0
mysql-connector-python==9.0.0
PyMySQL==1.1.1
sqlparse==0.5.1
uWSGI==2.0.26
